TP1/requirements.txt
pyodbc
python-dotenv
requests